pyproject.toml
[project]
name = "worldanvil-to-md"
version = "1.0.0"
description = "Convert World Anvil JSON exports to Obsidian markdown files."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx",
    "jinja2>=3.1.6",
    "pyyaml",
    "tqdm",
]

[tool.uv]
package = false
